+++ Java/spring-sprint-2/src/main/java/com/br/springsprint2/controle/ItensPedidosController.java
@@ -1,11 +1,22 @@
 package com.br.springsprint2.controle;
 
 import com.br.springsprint2.dominio.ItensPedido;
+import com.br.springsprint2.dominio.Petshop;
+import com.br.springsprint2.dominio.Produto;
+import com.br.springsprint2.dominio.Servico;
 import com.br.springsprint2.repositorio.ItensPedidosRepository;
+import com.br.springsprint2.repositorio.PetshopRepository;
+import com.br.springsprint2.repositorio.ProdutoRepository;
+import com.br.springsprint2.repositorio.ServicoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static org.springframework.http.ResponseEntity.noContent;
+import static org.springframework.http.ResponseEntity.status;
 
 @RestController
 @RequestMapping("itens")
@@ -13,25 +24,41 @@ public class ItensPedidosController {
     @Autowired
     private ItensPedidosRepository repository;
 
-    @PostMapping
-    public String createItens(@RequestBody ItensPedido novoItem) {
+    @Autowired
+    private ProdutoRepository prodRepository;
+
+    @Autowired
+    private ServicoRepository servRepository;
+
+    @PostMapping("/{fkServico}/{fkProduto}")
+    public ResponseEntity createItens(@RequestBody ItensPedido novoItem, @PathVariable int fkProduto, @PathVariable int fkServico) {
+        Produto produto = prodRepository.findById(fkProduto).get();
+        novoItem.setFkProduto(produto);
+        Servico servico = servRepository.findById(fkServico).get();
+        novoItem.setFkServico(servico);
         repository.save(novoItem);
-        return "Item cadastrado";
+        return status(HttpStatus.CREATED).build();
     }
 
     @GetMapping
-    public List<ItensPedido> getItens() {
-        return repository.findAll();
+    public ResponseEntity getItens() {
+        List<ItensPedido> lista = repository.findAll();
+        return lista.isEmpty() ? ResponseEntity.noContent().build() : ResponseEntity.ok().body(lista);
+
     }
 
     @GetMapping("/{id}")
-    public ItensPedido getItem(@PathVariable int id) {
-        return repository.findById(id).get();
+    public ResponseEntity getItem(@PathVariable int id) {
+        return ResponseEntity.of(repository.findById(id));
     }
 
     @DeleteMapping("/{id}")
-    public String deleteItem(@PathVariable int id) {
-        repository.deleteById(id);
-        return "Petshop excluido";
+    public ResponseEntity deleteItem(@PathVariable int id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200).build();
+        } else {
+            return ResponseEntity.status(404).build();
+        }
     }
 }
