+++ Java/spring-sprint-2/src/main/java/com/br/springsprint2/controle/ServicoController.java
@@ -5,9 +5,13 @@ import com.br.springsprint2.dominio.Servico;
 import com.br.springsprint2.repositorio.PetshopRepository;
 import com.br.springsprint2.repositorio.ServicoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static org.springframework.http.ResponseEntity.status;
 
 @RestController
 @RequestMapping("servico")
@@ -18,27 +22,32 @@ public class ServicoController {
     @Autowired
     private PetshopRepository petRepository;
 
-    @PostMapping("fkPetshop")
-    public String createServico(@RequestBody Servico novoServico, @PathVariable int fkPetshop) {
+    @PostMapping("{fkPetshop}")
+    public ResponseEntity createServico(@RequestBody Servico novoServico, @PathVariable int fkPetshop) {
         Petshop petshop = petRepository.findById(fkPetshop).get();
         novoServico.setFkPetShop(petshop);
         repository.save(novoServico);
-        return "Serviço cadastrado";
+        return status(HttpStatus.CREATED).build();
     }
 
     @GetMapping
-    public List<Servico> getServico() {
-        return repository.findAll();
+    public ResponseEntity getServico() {
+        List<Servico> lista = repository.findAll();
+        return lista.isEmpty() ? ResponseEntity.noContent().build() : ResponseEntity.ok().body(lista);
     }
 
     @GetMapping("/{id}")
-    public Servico getServico(@PathVariable int id) {
-        return repository.findById(id).get();
+    public ResponseEntity getServico(@PathVariable int id) {
+        return ResponseEntity.of(repository.findById(id));
     }
 
     @DeleteMapping("/{id}")
-    public String deleteServico(@PathVariable int id) {
-        repository.deleteById(id);
-        return "Petshop excluido";
+    public ResponseEntity deleteServico(@PathVariable int id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200).build();
+        } else {
+            return ResponseEntity.status(404).build();
+        }
     }
 }
