+++ Java/spring-sprint-2/src/main/java/com/br/springsprint2/controle/UsuarioController.java
@@ -1,13 +1,15 @@
 package com.br.springsprint2.controle;
 
-import com.br.springsprint2.dominio.Servico;
 import com.br.springsprint2.dominio.Usuario;
-import com.br.springsprint2.repositorio.ServicoRepository;
 import com.br.springsprint2.repositorio.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static org.springframework.http.ResponseEntity.status;
 
 @RestController
 @RequestMapping("usuarios")
@@ -16,24 +18,29 @@ public class UsuarioController {
     private UsuarioRepository repository;
 
     @PostMapping
-    public String createUsuario(@RequestBody Usuario novoUsuario) {
+    public ResponseEntity createUsuario(@RequestBody Usuario novoUsuario) {
         repository.save(novoUsuario);
-        return "Usuario cadastrado";
+        return status(HttpStatus.CREATED).build();
     }
 
     @GetMapping
-    public List<Usuario> getUsuarios() {
-        return repository.findAll();
+    public ResponseEntity getUsuarios() {
+        List<Usuario> lista = repository.findAll();
+        return lista.isEmpty() ? ResponseEntity.noContent().build() : ResponseEntity.ok().body(lista);
     }
 
     @GetMapping("/{id}")
-    public Usuario getUsuario(@PathVariable int id) {
-        return repository.findById(id).get();
+    public ResponseEntity getUsuario(@PathVariable int id) {
+        return ResponseEntity.of(repository.findById(id));
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUsuario(@PathVariable int id) {
-        repository.deleteById(id);
-        return "Usuario excluido";
+    public ResponseEntity deleteUsuario(@PathVariable int id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200).build();
+        } else {
+            return ResponseEntity.status(404).build();
+        }
     }
 }
